requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain
bs4
tiktoken
discord.py
aiohttp
psycopg2-binary
requests
streamlit